+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Format: YYYY-MM-DD — [type] short description (#issue, PR #)
 ## [Unreleased]
 - 2025-08-08 — [chore] Add PR/issue templates, CHANGELOG, BLOCKERS, TASKS; SSO helper path cleanup (#setup)
 - 2025-08-09 — [chore] Improve Soo script to better store the config certification. 
+- 2025-08-09 — [chore] Init Terraform and apply to create all the backend-s3-dynamodb-kms ( for storing the terraform and CI changes)